Identify all columns with potential issues

This function identifies all columns with potential issues, including all unique values, bijective columns, constant columns, and columns with missing values above a certain cutoff. It returns a list of identified columns for each category and can print the results to the console.

Usage

cols_identify_all(data, cutoff = 1, na.rm = FALSE, print = TRUE)

Arguments

data

The data frame

cutoff

The minimum cutoff for the proportion of missing values.

na.rm

Should missing values be removed?

print

Print the output or not.

Value

A named list with one element per check.

See also

Examples

cols_identify_all(airquality)
#> → No columns with all unique values.
#> → No bijective columns.
#> → No constant columns.
#> → No identified columns with missing proportion.
#> 
#> ── Checking if all values are unique in a column ───────────────────────────────
#> [1] NA
#> 
#> ── Checking if any two columns are bijective ───────────────────────────────────
#> list()
#> 
#> ── Checking if any column is all constant ──────────────────────────────────────
#> [1] NA
#> 
#> ── Checking if columns with preset cut-off in missing values ───────────────────
#> [1] NA